+++ app/src/main/java/com/freelancer/data/model/FirestoreRepository.java
@@ -245,7 +245,7 @@ public class FirestoreRepository {
             }
         });
     }
-    public void createJobReview(String comment, int numStars) {
+    public void createJobReview(String comment, float numStars) {
         Map<String, Object> jobReview = new HashMap<>();//Hashmap used to store key value pairs
 
         jobReview.put("Comment", comment);//storing appropriate data under the correct header
